Refactor of JS code
the json is now stored in a global variable which fixes the slight rounding errors that would happen when converting from imperial to metric and grabbing the rounded data from the html. This also means we no longer need a separate function for updating units so now the unit functions just use updateDisplay. Additional layout changes + deleting now unnecessary functions.